requirements.txt
Flask>=2.0.0
pandas>=1.3.0
openpyxl>=3.0.0
pystray>=0.19.0
Pillow>=9.0.0